Adds base camera class, stubs for quat camera.
