+++ RendAR/Sources/camera.cpp
@@ -1,0 +1,5 @@
+#include "camera.hpp"
+
+namespace RendAR {
+  
+}
